Add Todo Policy

+++ app/Http/Controllers/TodoController.php
@@ -30,6 +30,9 @@ class TodoController extends Controller
     }
 
     public function update(Todo $todo, TodoUpdateRequest $request) {
+
+        $this->authorize('update', $todo);
+
         $input = $request->validated();
         $todo->fill($input);
         $todo->save();
@@ -38,10 +41,16 @@ class TodoController extends Controller
     }
 
     public function destroy(Todo $todo) {
+
+        $this->authorize('destroy', $todo);
+
         $todo->delete();
     }
 
     public function show(Todo $todo) {
+
+        $this->authorize('view', $todo);
+
         // Get tasks Relation to pass in TodoResource resource
         $todo->load('tasks');
 
@@ -49,6 +58,9 @@ class TodoController extends Controller
     }
 
     public function addTask(Todo $todo, TodoTaskStoreRequest $request) {
+
+        $this->authorize('addTask', $todo);
+
         $input = $request->validated();
 
         $todoTask = $todo->tasks()->create($input);
